+++ back/src/bsds/resolvers/queries/__tests__/bsds.bsdasri.integration.ts
@@ -147,7 +147,7 @@ describe("Query.bsds.dasris base workflow", () => {
                 takenOverAt: new Date().toISOString() as any,
 
                 weight: { value: 99, isEstimate: false },
-
+                plates: ["TRANSPORTER-PLATE"],
                 packagings: [{ type: "FUT", quantity: 44, volume: 123 }],
 
                 acceptation: { status: WasteAcceptationStatus.ACCEPTED }