+++ Changelog.md
@@ -15,6 +15,8 @@ et le projet suit un schéma de versionning inspiré de [Calendar Versioning](ht
 
 #### :boom: Breaking changes
 
+- La plaque d'immatriculation est désormais obligatoire à la signature du transporteur (BSDD, BSFF, BSDASRI, BSDA) [PR 2528](https://github.com/MTES-MCT/trackdechets/pull/2528)
+
 #### :nail_care: Améliorations
 
 - BSFF - API : permettre de filtrer les BSFFs (query `bsffs`) sur le numéro SIRET du détenteur initial et sur le numéro de fiche d'intervention. [PR 2531](https://github.com/MTES-MCT/trackdechets/pull/2531)